requirements.txt
flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
sseclient-py==1.8.0
gunicorn==21.2.0 